refactor: decompose agent/tools.ts into focused tool modules
Split the 1,135-line createVaultTools() god function into 6 focused
modules under src/agent/tools/:
- types.ts: shared VaultToolDeps interface and textResult helper
- skill-validator.ts: validateSkillContent and related constants
- read-tools.ts: 8 read-only tools
- write-tools.ts: 9 write tools
- observability-tools.ts: 1 observability tool
- skill-tools.ts: 3 skill lifecycle tools

tools.ts is now 156 lines — assembles tool groups, exports MCP server
factories, and re-exports for backward compat. All 21 tools preserved,
VAULT_TOOL_COUNT unchanged.
